• Rename the Explore page to Browse. • Rename DNA Chat to Chat • Rename Overview Report to Analyze • Add a new Explore page which lets the user navigate to a random study page.
