+++ app/dna-chat/layout.tsx
@@ -1,14 +1,14 @@
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
-  title: "DNA Chat - Monadic DNA Explorer",
+  title: "Chat - Monadic DNA Explorer",
   description: "Ask private AI questions about your saved genetic results. Your DNA data never leaves your device.",
   keywords: ["DNA chat", "genetic AI", "private DNA analysis", "personal genomics AI", "DNA questions"],
   alternates: {
     canonical: "https://explorer.monadicdna.com/dna-chat",
   },
   openGraph: {
-    title: "DNA Chat - Monadic DNA Explorer",
+    title: "Chat - Monadic DNA Explorer",
     description: "Ask private AI questions about your saved genetic results. Your DNA data never leaves your device.",
     type: "website",
     url: "https://explorer.monadicdna.com/dna-chat",
@@ -17,7 +17,7 @@ export const metadata: Metadata = {
   },
   twitter: {
     card: "summary_large_image",
-    title: "DNA Chat - Monadic DNA Explorer",
+    title: "Chat - Monadic DNA Explorer",
     description: "Ask private AI questions about your saved genetic results. Your DNA data never leaves your device.",
     creator: "@MonadicDNA",
   },
