+++ app/explore/layout.tsx
@@ -1,14 +1,8 @@
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
-  title: "Explore GWAS Studies - Monadic DNA Explorer",
-  description: "Search and filter millions of genetic associations from the GWAS Catalog. Upload your DNA data to see personalized results.",
-  keywords: ["GWAS", "genetic studies", "DNA research", "genome-wide association", "genetic variants", "SNP analysis"],
-  openGraph: {
-    title: "Explore GWAS Studies - Monadic DNA Explorer",
-    description: "Search millions of genetic associations and analyze your DNA data",
-    type: "website",
-  },
+  title: "Explore - Monadic DNA Explorer",
+  description: "Discover a random study from the GWAS Catalog.",
 };
 
 export default function ExploreLayout({
